backend/pyproject.toml
[project]
name = "backend"
version = "0.1.0"
description = "Add your description here"
readme = "README.md"
requires-python = ">=3.13"
dependencies = [
    "cryptography>=46.0.3",
    "fastapi[all]>=0.126.0",
    "pymysql>=1.1.2",
    "python-dotenv>=1.2.1",
    "requests>=2.32.0",
    "sqlalchemy>=2.0.45",
    "uvicorn>=0.38.0",
]
